requirements.txt
# Core LLM + tools
openai>=1.40.0
mistralai>=1.2.0

# LangChain ecosystem
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0

# Vector store backend
faiss-cpu>=1.8.0

# Environment & web UI
python-dotenv>=1.0.1
gradio>=4.44.0
